refactor(server): extract the pure listing domain core
Move sorting, pagination, scope filtering, concept→game mapping, and the
discounted-classification allow-list into server/src/domain/listing.ts — a pure
module with no I/O imports, decoding via Schema.decodeUnknownSync. gamesService
now consumes the domain core; its infra/interface layer (Express, cache, env)
is rewritten to Effect in the next commit, where the server goes green again.
